# spec/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Business Watchdog core API
  version: 1.0.0
  description: |
    Core contracts for P0 and P1. This document is a specification, not a running API.
    HMAC canonicalization and transport limits are normative in TZ sections 11 and 27.
    UI CRUD, auth, billing, export and credential rotation must be added before their implementation.
    All monetary amounts are decimal strings in minor units. All timestamps are RFC3339 UTC.
servers:
  - url: https://watchdog.example.invalid
...
paths:
...
  /api/v1/tenants/context:
    get:
      tags: [Tenancy]
      operationId: tenantContext
      security:
        - {UserSession: []}
      responses:
        "200":
          description: Resolved tenant context
          content:
            application/json:
              schema:
                type: object
                properties:
                  active_tenant_id: {type: [string, "null"], format: uuid}
                  tenant_context_id: {type: [string, "null"], format: uuid}
                  tenant_context_source: {type: [string, "null"]}
        "401": {$ref: "#/components/responses/Problem"}
  /api/v1/stores:
    get:
      tags: [Stores]
      operationId: listStores
      security:
        - {UserSession: []}
      responses:
        "200":
          description: Stores of the active tenant with coverage
          content:
            application/json:
              schema:
                type: object
                required: [data, next_cursor]
                properties:
                  data:
                    type: array
                    items: {$ref: "#/components/schemas/Store"}
                  next_cursor: {type: [string, "null"]}
        "401": {$ref: "#/components/responses/Problem"}
    post:
      tags: [Stores]
      operationId: createStore
      security:
        - {UserSession: []}
      description: Owner/admin.
      parameters:
        - {$ref: "#/components/parameters/CsrfToken"}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [name, base_url, timezone, default_currency]
              properties:
                name: {type: string, minLength: 1, maxLength: 100}
                base_url:
                  type: string
                  format: uri
                  description: "Public HTTPS origin; private destinations, credentials, fragments and non-standard ports are rejected."
                timezone: {type: string}
                locale: {enum: [ru, en, de]}
                default_currency: {type: string, pattern: "^[A-Z]{3}$"}
      responses:
        "201":
          description: "Created (onboarding, unverified)"
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Store"}
        "401": {$ref: "#/components/responses/Problem"}
        "403": {$ref: "#/components/responses/Problem"}
        "422": {$ref: "#/components/responses/Problem"}
  /api/v1/stores/{store_id}:
    get:
      tags: [Stores]
      operationId: getStore
      security:
        - {UserSession: []}
      parameters:
        - name: store_id
          in: path
          required: true
          schema: {type: string, format: uuid}
      responses:
        "200":
          description: Store with coverage summary
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Store"}
          headers:
            ETag:
              description: Quoted version for If-Match
              schema: {type: string}
        "401": {$ref: "#/components/responses/Problem"}
        "404": {$ref: "#/components/responses/Problem"}
    patch:
      tags: [Stores]
      operationId: updateStore
      security:
        - {UserSession: []}
      description: Owner/admin; If-Match carries config_version. Activation and browser checks require a verified domain (422 store_not_verified).
      parameters:
        - name: store_id
          in: path
          required: true
          schema: {type: string, format: uuid}
        - {$ref: "#/components/parameters/IfMatch"}
        - {$ref: "#/components/parameters/CsrfToken"}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                name: {type: string, minLength: 1, maxLength: 100}
                base_url: {type: string, format: uri, description: Changing it clears verification and browser checks.}
                timezone: {type: string}
                locale: {enum: [ru, en, de]}
                default_currency: {type: string, pattern: "^[A-Z]{3}$"}
                status: {enum: [paused, active]}
                browser_enabled: {type: boolean}
                telemetry_enabled: {type: boolean}
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Store"}
          headers:
            ETag:
              description: Quoted version for If-Match
              schema: {type: string}
        "401": {$ref: "#/components/responses/Problem"}
        "403": {$ref: "#/components/responses/Problem"}
        "404": {$ref: "#/components/responses/Problem"}
        "409": {$ref: "#/components/responses/Problem"}
        "422": {$ref: "#/components/responses/Problem"}
        "428": {$ref: "#/components/responses/Problem"}
...
components:
  securitySchemes:
    IntegrationKeyId:
      type: apiKey
      in: header
      name: X-BW-Key-Id
      description: Key ID plus mandatory raw-body HMAC signature headers; key ID alone never authenticates.
    UserSession:
      type: apiKey
      in: cookie
      name: watchdog_session
      description: Same-origin Sanctum session with CSRF on mutations.
    WorkerToken:
      type: http
      scheme: bearer
      description: Scoped worker credential restricted to private ingress.
  parameters:
...
    IfMatch:
      name: If-Match
      in: header
      required: true
      description: Strong ETag containing quoted incident revision, for example "3".
      schema: {type: string, pattern: '^"[0-9]+"$'}
    CsrfToken:
      name: X-XSRF-TOKEN
      in: header
      required: true
      schema: {type: string}
...
  responses:
    Problem:
      description: Structured safe problem response; HTTP status determines retry policy.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Problem'
  schemas:
...
    Problem:
      type: object
      additionalProperties: false
      required: [code, message, request_id]
      properties:
        code: {type: string}
        message: {type: string}
        request_id: {type: string, format: uuid}
        errors: {type: array, items: {type: object}}
        retry_after_seconds: {type: integer, minimum: 0}
...
    StoreCoverage:
      type: object
      required: [connector, money, payment_attempts, browser_checks]
      properties:
        connector:
          type: object
          required: [state]
          properties:
            state: {enum: [not_connected, warming_up, fresh, stale, partial]}
            provider: {type: [string, "null"]}
            last_heartbeat_at: {type: [string, "null"], format: date-time}
            plugin_version: {type: [string, "null"]}
        money:
          type: object
          required: [state, providers]
          properties:
            state:
              enum: [provider_not_connected, store_not_connected, store_data_stale, reconciling]
              description: "Money is reconciled only with a connected independent provider and fresh store data; otherwise results are unknown, never \"fine\"."
            providers:
              type: array
              items:
                type: object
                required: [provider, status]
                properties:
                  provider: {type: string}
                  status: {type: string}
        payment_attempts:
          type: object
          required: [state]
          properties:
            state: {enum: [store_not_connected, store_data_stale, no_attempts_seen, no_recent_attempts, observing]}
            last_window_end_at: {type: [string, "null"], format: date-time}
        browser_checks:
          type: object
          required: [state]
          properties:
            state:
              enum:
                - store_not_verified
                - disabled
                - not_configured
                - scenario_disabled
                - scheduled
                - passing
                - failing
                - not_conclusive
            last_run_status: {type: [string, "null"]}
            last_run_error_code: {type: [string, "null"]}
            last_run_finished_at: {type: [string, "null"], format: date-time}
            next_due_at: {type: [string, "null"], format: date-time}
    Store:
      type: object
      required:
        - id
        - name
        - base_url
        - platform
        - timezone
        - locale
        - default_currency
        - status
        - verified_at
        - browser_enabled
        - telemetry_enabled
        - config_version
        - coverage
        - last_successful_check_at
        - active_incident_count
      properties:
        id: {type: string, format: uuid}
        name: {type: string}
        base_url: {type: string, format: uri}
        platform: {type: string}
        timezone: {type: string}
        locale: {enum: [ru, en, de]}
        default_currency: {type: string, pattern: "^[A-Z]{3}$"}
        status: {enum: [onboarding, active, paused, degraded, deleted]}
        verified_at: {type: [string, "null"], format: date-time}
        browser_enabled: {type: boolean}
        telemetry_enabled: {type: boolean}
        config_version: {type: integer, minimum: 1}
        coverage: {$ref: "#/components/schemas/StoreCoverage"}
        last_successful_check_at: {type: [string, "null"], format: date-time}
        active_incident_count: {type: integer, minimum: 0}
...